docs: update deploy instructions
We're noticing that the `main` branch is not always the appropriate source
for our online documentation, because it often contains updates to the
source docs that reflect changes that haven't been released yet. Therefore,
this commit proposes a new `docs-release` branch that is prepared ad hoc
before documentation is uploaded to our public website.

Closes #447