protocol/tx: add extensible tx hashing
This new package implements a change to transaction hashing that is more
flexible and futureproof. Transactions are converted to new data structures
that are intended eventually to replace the types in `protocol/bc`. This
transitional code is intended to protect against hash values changing when
we reach that point.

The spec for these changes is in
`docs/protocol/specifications/txgraph-draft.md`.

Still to come (in follow-on PRs): the code that integrates this new package
with the existing code in `protocol/bc`; and integration of
`txgraph-draft.md` into the spec proper.

Closes #458