sdk/ruby: fix path joining
The Connection class does not appropriately join URL path components
when both trailing and leading slashes occur in adjacent components.
This can occur when connecting to a real HSM server and attempting to
sign.

I chose a more brute-force approach for this fix, despite the existence
of something like URI::join. The latter has issues when adjacent paths
are both relative. The fix in this commit deals with this case
correctly.

Closes #502