core/leader: run lead func synchronously
This refactors the `core/leader` package and how it's used by `cmd/cored`:
* the `lead` closure is run synchronously within the `leader.Run` loop
* the global `IsLeading` value isn't updated until the `lead` closure
completes
* the`cmd/cored`'s `lead` closure calls `Recover` instead of pushing the
responsibility to the `core/fetch` and `core/generator` packages
* the `leader.Run` loop responsible for invoking `lead` is separated from
the loop responsible for detecting leadership changes

This ensures that all leadership-only routines and code paths are
guaranteed to have a recovered, current blockchain state. It also ensures
that if blockchain recovery takes longer than 10 seconds, a new leader
won't be demoted prematurely.

In a followup PR, we'll be able to remove
[this](https://github.com/chain/chain/blob/main/cmd/cored/main.go#L350-L368)
goroutine from `cmd/cored`, move the pin creating into the `lead` closure
and delete the `protocol.Chain.Ready` function.

Fixes #489.

Closes #497