core: don't initialize handler on first request
Rather than initializing the http handler on the first request, this
makes the creation of the handler an explicit step. To avoid confusion
with naming, the function that creates the http handler has been named
Handler, and the struct that used to be Handler is now API.

Additionally, action decoders has been changed from a map to a function.
The decoders should be available whether or not the handler has been
created.