docs/protocol: remove block witness from under the block ID
This makes block hash (aka "block ID") not subject to block witness
malleability. 

Rationale: turns out, when the block is agreed upon in a fully
decentralized manner (e.g. any BFT consensus protocol), not via a
assumed-to-be-alive generator, then nodes would require another round of
agreement on _which_ combination of public signatures to commit to. What's
worse, alternative combinations may be propagated quicker to other nodes
that must be then ready to re-assign witnesses when later blocks arrive to
keep on the same blockchain. 

Removing signatures from the chain of hashes eliminates this problem. Note
that other nodes consuming the chain of headers or full blocks are still
required to also receive and process witness data to authenticate the
chain. Each block's witness still authenticates the entire blockchain under
it.

Closes #343