cmd/cored: use latestVersion constant only for dev builds
This slightly simplifies the logic for initializing the
version string on startup. (It shouldn't be complicated!)
Notably, this means that the release version string is
derived entirely from the build tag, and it doesn't use
the latestVersion constant at all -- this way, release
version strings are consistently derived from a single
source of truth, the git tag.