protocol/patricia: restrict Contains to new protocol
The protocol defines the Merkle Patricia Tree as a set,
not a dictionary[1]. Our usage of this package fits the
definition, but this package's own interface doesn't
match yet.

This patch updates Contains to take only a single "item"
argument, which is used as both the key and the value.
In the "set" definition, there is no distinction between
keys and values -- the same bit string fills both roles.
We also remove ContainsKey since it is no longer necessary.

Future patches will also update Insert and Delete
analogously.

[1]: https://github.com/chain/chain/blob/d60cd85416/docs/protocol/specifications/data.md#merkle-patricia-tree