all: include chain-core-server-1.1.2 in all builds
This updates all installers to compile cored version 1.1.2. Additionally,
it updates contents of the root changelog on `main` to point to each
release's on-branch version, to avoid redundant commits to both main and
the release. This commit should be safe to cherry-pick onto the
`1.1-stable` whenever the `chain-core-server-1.1.2` tag is finalized.

Complete list of updates:

- cored `latestVersion` constant
- corectl `version` constant
- Docker image build script
- Mac installer and changelog
- Windows installer and changelog


Closes #706