docs: use release branches as source of production docs
Now that we have the 1.1-stable branch, we can simplify our
documentation workflow by simply generating and uploading
documentation from the tip of that branch. For the time being, it
doesn't seem like a separate docs tag is necessary.

This commit updates various pieces of documentation to account for
this simplified workflow.

Closes #715