cmd/slashland: generate a "rev id" for each commit on main
Right now, when someone clones the repo and builds using 'go install
chain/cmd/cored', the resulting binary will have some minimal version
information: the latestVersion constant containing the version string of
the last release we made. This comes in handy when someone is reporting
trouble and we ask them to run `cored -version` to investigate. But this is
only a little better than nothing; there are typically hundreds of commits
between releases. Adding a unique string to every single revision (on main)
will give us a much more precise idea of what code is being run.

We generate source files for our four main languages: Go, Java, JavaScript,
and Ruby. It won't be difficult to add others if we need them.

This PR only begins recording the revision ID. We'll incorporate it into
the version output in followup work.

Closes #558