all: use pg.DB
Most places in the code base take a `pg.DB` in their external
interfaces. Some take a `*sql.DB`. Update the remaining instances of
`*sql.DB` to `pg.DB`.

This was motivated by trying to call `accounts.NewManager` with a
`pg.DB` and haing to do a `db.(*sql.DB)` cast.