core/leader: add State enum
Replace IsLeader() with an enum that can be FOLLOWING, RECOVERING or
LEADING. RECOVERING is used when a process has been elected leader but
has not yet finished recovering the blockchain and becoming leader.

Use this more fine-grained status to enable the /info endpoint even when
RECOVERING. This fixes snapshot downloads so that the /info endpoint can
be used to track the progress of a downloading snapshot.