core/leader: add State enum
Replace IsLeader() with an enum that can be FOLLOWING, RECOVERING or
LEADING. RECOVERING is used when a process has been elected leader but
has not yet finished recovering the blockchain and becoming leader.

Use this more fine-grained status to enable the /info endpoint even when
RECOVERING. This fixes snapshot downloads so that the /info endpoint can
be used to track the progress of a downloading snapshot.

It's OK for a Recovering process to call forwardToLeader. It'll detect
that
it's trying to forward to itself and return a 503 status code.

I tested this locally by connecting to testnet through dashboard. I was
able to
watch the snapshot download progress in the lefthand bar.

Closes #776