core: fix processor dependencies
Fix a few block processor dependencies that were incorrectly missing or
unnecessary.

expire ACPs processor

The expire account control programs processor used to have a dependency
on the query package's transaction annotator and indexer. However, we no
longer use the account control programs for annotation. We only use the
account_utxo table. Instead, we just need to make sure that the account
indexer has successfully added the utxos to the account_utxo table
before we delete the control programs.

asset processor

A block processor can be run on multiple blocks at once. Previously, if
two consecutive blocks both contained transactions issuing a previously
unknown asset, there was a race condition for which value would be
recorded in the first_block_height column. We can fix this by always
updating with the lower block height.

annotated tx query processor

The query block processor is responsible for marking annotated outputs
as spent on every spend. Since a block processor can be run on multiple
blocks in parallel, the query block processor had a race condition. For
example, an output k can be created in block n-1 and spent in block n.
If the block processor processes block n first and then block n-1, the
processor would create the annotated output and never mark it as spent.
This would permanently corrupt balance queries. If we instead wait for
block n-1 to be processed by the query processor, we ensure that the
output has been created by the time we try to mark it as spent.


I tested this change locally against testnet, creating a few transactions
through the java SDK.

Closes #783