protocol/validation: implement entries-based validation
Rewrite block and transaction validation in terms of "entries."

This implements the spec changes that (at the moment) live on the
`txgraph-spec` branch. Reviewers, please evaluate this PR against the
documents there.

What used to be called the "well-formedness" check, a.k.a. "context-free
validation," is now simply "validation" and lives in `protocol/validation`.
Validation that depends on blockchain state is now called "applying" and
lives in `protocol`.

The old tangle of different entrypoints to (different phases of) validation
has hopefully been simplified.

The blockchain state's "issuance memory" is replaced by a "nonce set."

Closes #788