cmd/slashland: fix logic to avoid empty commits
Running 'git commit' (without --allow-empty) will fail
when the file contents haven't changed, regardless of
the files' time stamps. We want to detect this situation
ahead of time and skip committing when the tree is
clean.

Unfortunately, diff-index, which is what we use to check
for a clean tree, *does* consider timestamps when
computing the diff; fortunately, running 'git add' fixes
this and causes the behavior of diff-index to match what
commit looks for in its "nothing to commit" error
message.

So now we explicitly add the files before checking to
see if the tree is clean.

This also gives us enough confidence to remove the
--allow-empty check on commit. It would be preferable
for chainbot simply to fail, alerting us that there's a
problem, rather than producing an empty commit and
continuing.

Fixes #830.

Closes #839