cmd/cored: configure root CA pool
By default, package crypto/tls uses the operating
system's set of global root CAs. This is unsuitable for
a typical production deployment, where the opeator will
run its own CA and provision its own certs.

This change provides a new environment variable:

ROOT_CA_CERTS=certs.pem
This variable holds a filename of a file containing a list
of PEM-encoded root X.509 certificates.

This is a back port of a change made in #684 with
updates to the logging call sites.

Closes #862