docker/ci: optimize image build times
This commit addresses #890 and #891. Previously, the image
building process involved compiling the protobuf compiler from
src. This takes a very long time, but was necessary due to glibc
incompatibilities between the official binaries and CentOS 6.

The artifacts of the compilation have been extracted from the
image and placed on s3 for download. This commit also separates
the image's install scripts from its test scripts and only loads the
former into the Dockerfile. This will allow updates to tests without
having to rebuild the image.