bin/wait-and-configure: new utility script
`wait-and-configure` waits until a local Chain Core daemon is ready, and
then attempts to configure it. You can use this script as part of a
quick reset-and-configure loop while iterating on code. For example,
some of our teammates use a Make target that executes the equivalent of
the following:

```
  go install $(CHAIN)/cmd/corectl
  go install $(REPO)/cmd/cored
  -dropdb core
  createdb core
  rm -rf ~/Library/Application\ Support/Chain\ Core/raft
  bin/wait-and-configure &
  cored
```

Previously, `corectl config-generator` could be used in place of
`wait-and-configure` for fast iteration loops, but since it now requires a
running instance of Chain Core, we need something that can run in the
background.

Closes #914