core/query: add `block_transactions_count` to tx
Add a `block_transactions_count` to annotated transactions.

If we wanted to backfill, we would need to not only backfill the
`block_tx_count` column but also add the element to the json in the `data`
column. Couple options:

1. Write a Go migration to backfill and then SET NOT NULL on the
`block_tx_count` column.
2. If we required PostgreSQL 9.5.1, we could write some intense SQL to
count the number of  annotated_txs at each block height and use the `||`
operator to merge the new json element into the `data` column's json blob.
This wouldn't respect the field ordering for the `block_transactions_count`
field in backfilled annotated transaction.

Fix #921.

Closes #927