core: trust the current process cert
A Chain Core TLS config is used by cored peers to talk
to each other, and by corectl to talk to cored. All
those processes have the same identity, so now we
automatically trust the local cert, with the expectation
that the peer will also be using it. This makes
authn misconfiguation impossible as long as the cert and
key files are loaded successfully.