protocol/bc/legacy: avoid erroring in MapTx
Currently we map legacy transactions into txgraph transactions when
deserializing them. Because MapTx was erroring on nonce-less issuances,
submitting an invalid transaction w/ a nonce-less issuance would error
during unmarshalling before reaching validation logic. The error
formatter would format this as an internal server error or a bad request
depending on the endpoint.

AFAICT, this is the only error condition in MapTx right now. The rest of
the error conditions seem like they'd only be hit if there was a bug in
our own code. @bobg does that seem right?

Should `MapTx`'s signature change to not return an error?