sdk/java: new Client.Builder constructor
The new constructor takes a Client object as an argument. The private
members of the Client are defensively copied into the resulting builder.
To facilitate this change, a private OkHttpClient member was added to
the builder (no setter method was provided). A new builder method was
also added to allow a user to set a list of URLs.

The impetus for this change was a bug I found when adding client cert
authn to the SDK. Our current method of creating a Client for the MockHsm
(`getSignerClient`) does not persist any configurations set on the httpClient.
The user can get around this by creating a separate Client with the appended
"/mockhsm" string but our provided mechanism is broken as is.