core: add new policy "public," new guard type "any"
We are going to gate the `/dashboard` routes behind the auth handlers
shortly. When we do that, we need to allow them to be accessible by any
client. 

(Dashboard will still have to make a request to cored to actually populate
the pages, and cored may respond with an unauth'd error, in which case
dashboard will show the "enter access token page." This PR allows dashboard
to show that page, instead of returning a 40X to the browser.) 

Closes #999