cmd/cored: avoid spurious 404 responses
The chain/core API object depends on reading a config
object during startup, and reading the config object
depends on reading from raft storage, which in turn
depends on handling HTTP requests. We previously handled
this by adding the API handler to the mux after setup.
This caused us to respond 404 to any non-raft request
received during the brief startup time.

Now, we install a handler in the mux immediately, but
synchronized with the API handler, so it blocks until
the API handler is ready. This should provide the
correct behavior for all HTTP responses at all times.

In the future, we might investigate lazily loading the
config object to avoid this awkward data dependency
sequence altogether.

Fixes #1053.

Closes #1058