cmd/cored: enforce TLS without build tag plain_http
With this change, TLS will be required by default, and can be made optional
with build tag plain_http. Our intention is to require TLS internally
during development (hence why the default build configuration is to require
it) and for production, but not to require TLS for desktop builds (Mac OS X
and Windows).

This change also unconditionally sets the secureheader field HTTPSRedirect
to false. Chain Core never listens on both TLS and non-TLS ports
simultaneously, so it is pointless to ever set this field to true. Doing so
was a longstanding mistake, which we might as well fix now.

Future work: consider renaming plain_http to plain_http_ok to better
reflect that TLS is optional (not prohibited) in that configuration.

Closes #1059