core: use Chain.State to retrieve latest bc state
Previously, core/generator and core/fetch had their own local pointers
to the latest block and latest snapshot. We already need to store and
expose blockchain state on the Chain struct for the /sign-block RPC.
To address #519 without too invasive of changes, I think we'll need to
pull the state from the shared Chain struct.

Closes #1073