database/raft: create the allowed members list
When a new cored joins a raft cluster, there must be an associated grant
stored in that cluster in order for this new cored to talk to the other
members of the cluster. 

We'd like the existence of this grant to be a prerequisite for a new member
joining the cluster. Similarly, we'd like it to be impossible to delete a
grant that is associated with a cluster member, unless that member is also
removed from the cluster. In order to keep track of this, we're creating a
list of "allowed members"--that is, addresses associated with nodes that
are allowed to become members. 

This commit creates that list, but doesn't hook it up to the rest of the
code managing cluster membership. 

Closes #1082