database/raft: add useTLS config option
The standard library's `net/http` package modifies the http client's
transport field when requests are made. Therefore it's not safe to check if
we should use TLS by checking the transport. This PR adds a `useTLS` flag
to the raft service; in the future, the raft service may be configured by
passing in run options, a la cored. 

Closes #1097