database/raft: fix WAL snapshot after joining a cluster
We need to save the initial snapshot to the WAL
immediately when joining a cluster. Failing to do this
means subsequent entries are not saved until the next
snapshot, causing a crash on restart.

In the vendored raft package, on startup, the read
snapshot serves as the base index for the slice of
entries. The slice of entries is initially empty, and
entries are appended as they're read. Entries with an
index greater than the currently read entry are sliced
off.

The crash happens there. If we fail to save the initial
snapshot, then the base index will be 0 and the first
entry record has a nonzero index, causing an
out-of-bounds slice operation.