net/http/authn: return 401 when loopback_auth is off
This is a temporary workaround until we can more
carefully plan how to treat authn vs authz in various
clients such as Dashboard. Currently, we want to return
a 401 to indicate that the user needs to supply some
credentials, but we want to avoid returning that error
for requests to known-public resources (/dashboard/).

In the future, we should avoid ever returning 401,
internally treating authn as a set of orthogonal
credentials, which may legitimately be empty. (An empty
set of credentials is still useful when accessing public
resources.) Clients should then detect the "needs to
prompt for credentials" situation by requesting access
to a known lowest-common-denominator resource (most
likely /info) and looking for a 403 response to that
request.

Closes #1121