net/http/authn: explicitly verify client certs
The std library package crypto/tls provides several
options for ClientAuthType; see
https://godoc.org/crypto/tls#ClientAuthType. None of
these do what we want, which is to request a client
cert, and validate it if given, but to treat an invalid
cert the same as a missing client cert: the connection
should be permitted but the cert should not be used in
authz.

We can fix this by using RequestClientCert and calling
Verify ourselves in the authn package. This also
requires plumbing to give authn access to the cert pool.