database/sinkdb/sinkdbtest: new package
Add package with test utilities for creating test sinkdb instances.
This will also prevent tests in separate packages from interacting.

I played with doing cleanup in a
[finalizer](https://golang.org/pkg/runtime/#SetFinalizer) but there's no
guarantee
that the finalizer will run if the process exits shortly after. Maybe
that doesn't matter though and the simpler interface is worth it?

Closes #1197