core: set policies for all /debug/pprof/ endpoints
The handler installed at the path /debug/pprof/ serves the heap and
blocking profiles at /debug/pprof/heap and /debug/pprof/block. The
handler was installed at /debug/pprof/ but the policies were defined on
the path /debug/pprof, so these profiles were inaccessible.

Alternatively, if we want to allow access to all future profiles that
might be exposed by the pprof package, we could use a single authz entry
for the /debug/pprof/ path.