docs: clarify backward compatibility rules for SDKs
This commit refines our versioning language such that
backward compatibility guarantees only hold between server packages and
client packages of the previous minor version. In particular, this
is intended to eliminate any interpretation that holds that new SDKs
should be backward-compatible with application code written for a
previous SDK version.

The goal of this change is to allow us to pursue a more aggressive
release roadmap, and to eliminate versioning kludges in the SDK. It
allows us to introduce new usage patterns (potentially breaking changes)
in consecutive minor versions of the SDK without eliminating a smooth
upgrade path for application code.

Closes #1225