net/raft: campaign immediately in Init
Immediately campaign for leader when initializing a new cluster. This
should speed up tests. We'll no longer need to wait electionTick * tickDur
before having a usable raft node.

If we convert to raft.RawNode at some point, we could avoid the
less-than-ideal synchronous Ready channel wait. Because StartNode
documents that it appends ConfChangeAddNode entries for each peer
to the initial log, I think we are guaranteed that the channel wait will
not block indefinitely. We might not be guaranteed that it'll actually
contain all the initial config changes, but in that case Campaign will
fail silently and the node will campaign after electionTick ticks.

See coreos/etcd#7985.

Closes #1242