net/raft: crash process on unrecoverable errors
cored cannot operate without an active raft service. Chain Core
leadership is maintained Postgres, so crashing the process ensures
that the leader will abdicate if its raft service fails.

Since runUpdates runs forever, there is now no way for a raft.Service
to be marked as 'done'. If we want to add a way to gracefully shutdown
a raft service in the future, we can add back the donec channel and
related logic.

Closes #1255