docs/future: new directory
`docs/future` should be the home for any documentation update that is not
yet considered releasable. This includes:

- updates that will be included in the next minor version (x.y) release.
These updates may be merged to `main`. When appropriate, `docs/future` will
be copied and renamed with a proper release number.
- updates that will be included in some undetermined future date. These
updates should live in a separate branch, and not be merged into `main`.

The contents of `docs/future` in this commit is just a copy of the
`docs/1.2` directory.

Closes #1308