vendor: add google/snappy and introduce patches
Add snappy, which is a build dependency for rocksdb. 

Per conversation with @ryandotsmith,  also create the `vendor/patches.md`
file, which we will use to track small changes (i.e., changes not worth
creating a fork for) to our vendored dependencies. In the future, we will
likely have a more sophisticated patch system. 

Closes #1316