net/raft: save snapshot position first
Save the snapshot position to the WAL before saving the snapshot data to
disk. On boot, we open the WAL at the most recently saved snapshot's
position. If we don't save the position to the WAL first, we might open
the WAL at a snapshot position that was never saved to the WAL.

See coreos/etcd#8082, coreos/etcd#8088.

Closes #1312