core: add /evict rpc endpoint
Add an evict command to corectl to remove a node from a Chain Core
cluster.

On startup in cmd/cored, call sinkdb.DB.Ping to perform an empty write
to test the raft configuration. This helps handle the scenario where an
evicted node boots up without realizing it's been evicted. The other
nodes won't necessarily replicate the eviction entry to the node if
they've already moved on and forgotten the evicted node's address.

Addresses part of #1118.

Closes #1269